CAP ASP 2018 prep - created results file
this is an excel formatted results spreadsheet. with xlsx formatting we can create tabs, save formulas etc. all the usual excel stuff
lets use this file to do all of our calculations (and potentially graphing, though due to the issues with time formatting it may be easier to use R for graphics)